% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_share.R
\name{top_share}
\alias{top_share}
\title{Top Income Share}
\usage{
top_share(x, p = 0.1)
}
\arguments{
\item{x}{Numeric vector of incomes.}

\item{p}{Proportion of the top income group (default is 0.1 for top 10\%).}
}
\value{
Numeric value between 0 and 1 representing the share.
}
\description{
Calculates the share of total income held by the top p proportion.
}
\examples{
top_share(c(1000, 2000, 3000, 4000, 5000), p = 0.2)
}
